% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlatToXY.R
\name{bearingToXY}
\alias{bearingToXY}
\title{Convert a bearing (in degrees) and velocity (in knots) to north and east 
velocity (in ft / s).}
\usage{
bearingToXY(bearing, speed)
}
\arguments{
\item{bearing}{A numeric vector giving the instantaneous direction of the 
velocity in degrees.}

\item{speed}{A numeric vector giving the instantaneous magnitude of the 
velocity in knots.}
}
\value{
A n x 2 matrix giving the north/south and east/west components of the
  velocity, in feet/s.
}
\description{
Convert a bearing (in degrees) and velocity (in knots) to north and east 
velocity (in ft / s).
}

