% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSLoWC.R
\name{calculateSLoWC}
\alias{calculateSLoWC}
\title{Calculate the "severity loss of Well Clear" (SLoWC) metric, described in RTCA
SC-228 Closed-Loop Metrics White Paper.}
\usage{
calculateSLoWC(trajectory1, trajectory2)
}
\arguments{
\item{trajectory1}{A \code{flighttrajectory} object corresponding to the
first aircraft.}

\item{trajectory2}{A \code{flighttrajectory} object corresponding to the
second aircraft.}
}
\value{
The numeric vector giving the SLoWC metric. Values lie in the range 
  [0, 100]. A SLoWC of 0 indicates Well Clear, while a value of 100 
  corresponds to "full penetration" (i.e., a collision).
}
\description{
Calculate the "severity loss of Well Clear" (SLoWC) metric, described in RTCA
SC-228 Closed-Loop Metrics White Paper.
}
\details{
Note that the RTCA definition of Well Clear is undergoing revision. 
  This code is based on the SLoWC formulation by Ethan Pratt and Jacob Kay as
  implemented in a MATLAB script by Ethan Pratt dated 2016-04-18.
}

