% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightconflicts.R
\docType{package}
\name{flightconflicts}
\alias{flightconflicts}
\alias{flightconflicts-package}
\title{flightconflicts: Analyze conflicts between aircraft}
\description{
This package contains tools to analyze aircraft trajectories and identify
loss of well-clear (LoWC) and near mid-air collisions (NMAC).
}

