% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyViolations.R
\name{identifyLoWC}
\alias{identifyLoWC}
\title{Determine whether loss of Well Clear (LoWC) has occurred at each time point.}
\usage{
identifyLoWC(trajectory1, trajectory2)
}
\arguments{
\item{trajectory1}{A \code{flighttrajectory} object corresponding to the 
first aircraft.}

\item{trajectory2}{A \code{flighttrajectory} object corresponding to the 
second aircraft.}
}
\value{
The logical vector indicating whether Well Clear has been violated at
  each time point.
}
\description{
Determine whether loss of Well Clear (LoWC) has occurred at each time point.
}
\details{
This code relies on \code{\link{calculateSLoWC}}, see its
  documentation for details.
}

