% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNMAC.R
\name{identifyNMAC}
\alias{identifyNMAC}
\title{Determine whether an NMAC (near mid air collision) has occurred at each time 
point.}
\usage{
identifyNMAC(trajectory1, trajectory2)
}
\arguments{
\item{trajectory1}{A \code{flighttrajectory} object corresponding to the
first aircraft.}

\item{trajectory2}{A \code{flighttrajectory} object corresponding to the
second aircraft.}
}
\value{
The logical vector indicating whether NMAC criteria are met at each 
  time point.
}
\description{
Determine whether an NMAC (near mid air collision) has occurred at each time 
point.
}
\details{
NMAC is a cylinder +/- 100 ft above and below the ownship with a 
  radius of 500 ft.
}

