% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryToXYZ.R
\name{trajectoryToXYZ}
\alias{trajectoryToXYZ}
\title{Convert trajectories to Easting/Northing representation.}
\usage{
trajectoryToXYZ(trajectory, origin)
}
\arguments{
\item{trajectory}{A flighttrajectory object.}

\item{origin}{A length 2 numeric vector giving the origin lon/lat for the
conversion.}
}
\value{
A flattrajectory object.
}
\description{
Convert trajectories to Easting/Northing representation.
}

