% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTrajectories.R
\name{checkTrajectories}
\alias{checkTrajectories}
\title{Check that both arguments are trajectories with the same timecourse.}
\usage{
checkTrajectories(trajectory1, trajectory2)
}
\arguments{
\item{trajectory1}{An object to test.}

\item{trajectory2}{Another object to test.}
}
\value{
TRUE if everything works. Raises an error otherwise.
}
\description{
Check that both arguments are trajectories with the same timecourse.
}

